Updated lifecycle script for easier maintenance
Updated bash script so that it is more resilient to changes. It should install the connector in all tensorflow, mxnet, pytorch and python environments instead of a select hardcoded environments. I have also pinned this CFT version to the SF connector version 2.0.3 to also minimize risk of dependency issues down the road.